% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailzoekR.R
\name{basisgegevens_instellingen}
\alias{basisgegevens_instellingen}
\title{Haalt organisaties.csv uit wekelijkse DUO data}
\usage{
basisgegevens_instellingen()
}
\description{
Haalt organisaties.csv uit wekelijkse DUO data
}
